Fixing eBookDefinition for eBookManager
git-tfs-id: [https://tfsnpt.int.thomson.com:443/tfs/Cobalt_Collection/]$/Cobalt eBook Builder/Development;C192385

+++ ebookGeneratorCore/src/com/thomsonreuters/uscl/ereader/orchestrate/core/dao/CoreDao.java
@@ -39,7 +39,7 @@ public interface CoreDao {
 	 * Returns a count of the current book definitions.
 	 * @return an integer
 	 */
-	public int countNumberOfBookDefinitions();
+	public long countNumberOfBookDefinitions();
 	
 	
 	/**
